simplify(mcp): drop dead refs + share core test helper
irc-server.ts:
- remove MAX_CHUNK_BODY import (unused)
- remove RecvBuf.firstSeen (set, never read; flushTimer drives flush)
- remove multilineMaxBytes (parsed/logged, never gates anything)

test/helpers: factor mcp.ts and mcp-inprocess.ts onto a shared
mcp-core.ts (ChannelNotification, wireMcpClient, pollUntilIrcReady).
Both helpers shrink to transport bring-up + teardown.

Per #63 simplify pass, with @lead-pm approval. In-process coverage
path preserved (irc-server.ts still surfaced to bun's instrumentation).

constants.ts inline considered; declined as too marginal (one constant
shared between irc-lib.ts, irc-server.ts, and two test files).

## Human-Claude Interaction Log

### Human prompts (VERBATIM):
1. (Spawn prompt — reviewer-63 doing /simplify holistic pass on #63.
   Workflow per spawn: read issue, plan in #issue-63, await lead-pm
   approval, produce cleanup PR + issues.)
   → Claude: read repo, posted plan in #issue-63, lead-pm approved
     items 1+2 as cleanup PR, 3 as comment on #59, 4 as new issue,
     5 as PR-description note. This commit is items 1+2+5.
